V2

+++ src/main/java/io/daut/crypto/wealth/inbound/FortuneFile.java
@@ -4,6 +4,7 @@ import io.daut.crypto.wealth.core.Currency;
 import io.daut.crypto.wealth.core.Money;
 
 import java.io.BufferedReader;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -24,6 +25,6 @@ public class FortuneFile {
     private static Money toMoney(String[] line) {
         final Currency currency = Currency.from(line[0]);
         final double value = Double.parseDouble(line[1]);
-        return new Money(currency, value);
+        return new Money(currency, BigDecimal.valueOf(value));
     }
 }
